flyio/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
black==22.10.0
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.0
frozenlist==1.3.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.21.6
pandas==1.3.5
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.4
pyee==9.0.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
slackclient==2.9.4
slackeventsapi==3.0.1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
